feat: support symbolic linked pipelines to avoid repetition (#580)
## why 

While using kpops and trying to reduce the amount of copy/pasting
required to configure multi-tenant configuration I found that _symbolic
links_ would cause kpops execution to break.

## what 

Replaces existing `Path.blob` with `blob.blob` and resolves the real
file location so no other changes are required for it to be compatible
with existing logic.

> [!NOTE]
> In python 3.13 symbolic links are supported
[docs.python.org/3.13#pathlib.Path.glob](https://docs.python.org/3.13/library/pathlib.html#pathlib.Path.glob),
probably it would be as easy as follows:
> `yield from sorted(pipeline_path.glob(f"**/{PIPELINE_YAML}",
recurse_symlinks=True))`